pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core"
]

[tool.poetry]
authors = ["Bernardo Couto <bernardocouto@icloud.com>"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Database"
]
description = "Databricks SQL framework, easy to learn, fast to code, ready for production."
homepage = "https://github.com/bernardocouto/databricks-sql"
license = "Apache 2-0"
name = "databricks-sql"
readme = "README.md"
repository = "https://github.com/bernardocouto/databricks-sql"
version = "1.0.0"

[tool.poetry.dependencies]
databricks-sql-connector = "2.2.1"
pystache = "0.6.0"
python = "^3.8"

[tool.poetry.dev-dependencies]
pre-commit = "2.20.0"
